chore(popup): modularize replacement suggest component
